requirements.txt
requests>=2.31.0
pillow>=10.0.0
opencv-python>=4.8.0
pandas>=2.0.0
python-dotenv>=1.0.0
flask>=2.3.0
werkzeug>=2.3.0
pytesseract>=0.3.10
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
schedule>=1.2.0
